UPSTREAM: tools/bootconfig: Show the number of bootconfig nodes
Show the number of bootconfig nodes when applying new bootconfig to
initrd.

Since there are limitations of bootconfig not only in its filesize,
but also the number of nodes, the number should be shown when applying
so that user can get the feeling of scale of current bootconfig.

Link: http://lkml.kernel.org/r/158091061337.27924.10886706631693823982.stgit@devnote2

Signed-off-by: Masami Hiramatsu <mhiramat@kernel.org>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
(cherry picked from commit 0f0d0a77ccbdc6bfe51c0d184f92d99ebcc2e809)
Bug: 157247714
Signed-off-by: Alistair Delva <adelva@google.com>
Change-Id: I59f486c6cd7c9e2b601afe538533c60d9a57a39c
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>