UPSTREAM: lib/bootconfig: Add override operator support
Add the value override operator (":=") support to the bootconfig.

This value override operator will be useful for the bootloaders
which will only update the existing bootconfig according to the
bootloader boot options.

Without this override operator, the bootloader needs to parse
the existing bootconfig and update it. However, with this
assignment, it can just append the updated (partial) bootconfig
text at the tail of existing one without parsing it.
(Of course, it must update the size, checksum and magic,
 but that will be done easily)

Link: https://lkml.kernel.org/r/159482882954.126704.16209517125614438640.stgit@devnote2

Signed-off-by: Masami Hiramatsu <mhiramat@kernel.org>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
(cherry picked from commit a2de2f86ae3831736dc906f9559b600f186403fe)
Signed-off-by: Devin Moore <devinmoore@google.com>
Change-Id: I328e39ffc756449e73882438a4adff79aa7b3e45
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>