msm: kgsl: Move SW ringbuffer support to target specific
Make nearly all of the software ringbuffer support target specific.
This allows us to implement specific targeted workarounds and gets rid of
some of the gpudev generic functions that are only implemented for one or
two targets. This does add a lot of code but in the end it makes it easier
for us to implement new generations without having to worry about breaking
older generations that still might be actively worked on.

Change-Id: Ic0dedbad87500aeb05c72c15d1bcf133241b4e1a
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Raghu Ananya Arabolu <rarabolu@codeaurora.org>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>