remoteproc: qcom: Set state of rproc to crashed before panicking
When the client crashes when recovery is disabled we call a kernel panic.
Since we will be trying to shutdown the remoteproc when it has already
crashed we will be waiting unecessarily for the remoteproc to respond to
the shutdown request. This change sets the state of the remoteproc to
crashed.

Change-Id: I66ef24925fce4d12c64693a8870edb9df3e2c2af
Signed-off-by: Siddharth Gupta <sidgup@codeaurora.org>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>