remoteproc: qcom: pas: Unlock metadata after auth and reset
This change adds a new API to mdt loader for remoteprocs that do not free
the metadata that is passed to the secure world. This is required as the
secure world locks the metadata buffer to ensure that no modifications to
the firmware can be made. The metadata buffer is unlocked once the auth
and reset call succeeds, after which the HLOS can free the metadata.

Change-Id: I078d6a19df98590c6cc38a70344c58598bb6ecf8
Signed-off-by: Siddharth Gupta <sidgup@codeaurora.org>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>