ARM: dts: msm: Remove redundant/unused secure display CMA region for yupik
The secure display region was kept around during the transition phase
from ION to dma-buf heaps. Now that the transition is complete, ION
has been removed, and the secure display region was absorbed into
the non-secure display region, we can remove the now unused region,
and stop including ION devicetree file.

Also, it has been observed that the kernel reorders the CMA regions
prior to allocating the memory for each region. This reordering can
cause smaller CMA regions to be allocated before bigger CMA regions,
resulting in failure to allocate the bigger CMA regions, as the memory
in the lower 4 GB becomes too fragmented. To mitigate this, assign a
region in memory to the non-secure CMA display region.

Change-Id: I27ca5b7115abd4620ae06ef588629f5fe19b020a
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>