dma-heap: qcom: Support per-CMA heap alignment
add_memory_subsection()/remove_memory_subsection() require the base and
size of the memory region to be aligned to 2Mb. These APIs are invoked
when transferring memory to SVM, and currently such allocations come
solely from the display heap.

This change does not attempt to handle the case where a memory region
smaller than 2MB is transferred to SVM.

Change-Id: I8296b5e5d9bc9bbf32b6a4094b8290e23b6e4354
Signed-off-by: Patrick Daly <pdaly@codeaurora.org>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>