power: supply: qti_battery_charger: Handle BC_SHIP_MODE_REQ_SET response
During shutdown when configuring ship mode based on user request,
battery charger driver sends BC_SHIP_MODE_REQ_SET message to the
charger firmware to do configuration before the shutdown. However
the response from the charger firmware is not handled and hence a
timeout is observed in the logs. Fix it.

Fixes: 1df95fc50926 ("power: supply: qti_battery_charger: Enable Ship mode shutdown")
Change-Id: I9149a256e4af40c3dcf2f0019f27e711a97f4493
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>