remoteproc: qcom: Add APIs to provide early notifications
Devices which cannot tolerate latencies in notification after a
remoteproc crash (such as PCIe), need a way to be notified in a
fraction of a second to ensure they meet the QoS requirements.
This change adds APIs to provide a way for such device drivers
to register for early notifications.

Change-Id: Ieee0c3868962fc825407d7292e4be9f10248c864
Signed-off-by: Siddharth Gupta <sidgup@codeaurora.org>
Signed-off-by: Mukesh Ojha <quic_mojha@quicinc.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>