remoteproc: qcom: Register SSR subdevice with the q6v5 IRQ driver
To allow for early notifications to be sent to the SSR clients the q6v5
driver needs a way to invoke the SSR subdevice for the remoteproc. This
change adds an API in the q6v5 IRQ layer to optionally register the SSR
subdevice, and registers the SSR subdevice in the PAS driver.

Change-Id: I450d262575538b6b2c2e038a41fd0cca234bc3b2
Signed-off-by: Siddharth Gupta <sidgup@codeaurora.org>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>